requirements.txt
aiohttp==1.0.5
async-timeout==2.0.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
discord.py==0.16.12
idna==2.6
multidict==4.1.0
pip==9.0.1
pycparser==2.18
PyNaCl==1.2.1
requests==2.18.4
setuptools==28.8.0
six==1.11.0
urllib3==1.22
websockets==3.4
youtube-dl==2018.4.3
